Return the exit code of the subprocess
When running asciinema with --command, it is useful to be able to
inspect the exit code of the child process. This allows for asciinema to
be used in automation/scripts "as if" it were the executed command.

A specific use case for this is using asciinema to record the terminal
output of running pty tests, such as testing vim.

TODO: Should this be protected by an option? This is a strict change in
behaviour.